backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0
pydantic==2.6.4
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.1.2
redis==5.0.1
celery==5.3.6
httpx==0.26.0
boto3==1.34.11
minio==7.2.0
ffmpeg-python==0.2.0
python-dotenv==1.0.0
loguru==0.7.2
orjson==3.9.10
jinja2==3.1.2
itsdangerous==2.1.2
pytest==8.2.2
PyJWT==2.8.0
cryptography==41.0.7
opentelemetry-sdk==1.23.0
opentelemetry-api==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-exporter-otlp==1.23.0
